Late Night Refactor
Made sure to look at the changes to get a good idea of where to go next, and reformatted some of the lines and names just to keep things consistent for when I have to move stuff in and out of the server.

Also I think this repo is haunted? I saw a line for a file pop in and out really quickly, but nothing else happened so....